Refactored road network creation
