+++ src/OpenSage.Game/Terrain/RoadTopology.cs
@@ -39,133 +39,6 @@ namespace OpenSage.Terrain
             }
             return node;
         }
-
-        public List<RoadNetwork> BuildNetworks()
-        {
-            var networks = new List<RoadNetwork>();
-
-            var edgeSegments = Edges.ToDictionary(e => e, e => new AngledRoadSegment(e.Start.Position, e.End.Position));
-            var nodeSegments = new Dictionary<RoadTopologyNode, IRoadSegment>();
-
-            foreach (var edge in Edges)
-            {
-                var edgeSegment = edgeSegments[edge];
-
-                void Connect(RoadTopologyNode node, in Vector3 direction)
-                {
-                    foreach (var connectedEdge in node.Edges)
-                    {
-                        if (connectedEdge == edge || connectedEdge.Template != edge.Template)
-                        {
-                            continue;
-                        }
-
-                        var connectedEdgeSegment = edgeSegments[connectedEdge];
-
-                        if (connectedEdge.Start.Position == node.Position)
-                        {
-                            connectedEdgeSegment.Start.ConnectTo(edgeSegment, direction);
-                        }
-                        else
-                        {
-                            connectedEdgeSegment.End.ConnectTo(edgeSegment, direction);
-                        }
-                    }
-                }
-
-                Connect(edge.Start, edge.Start.Position - edge.End.Position);
-                Connect(edge.End, edge.End.Position - edge.Start.Position);
-            }
-
-            foreach (var node in Nodes)
-            {
-                foreach (var edgesPerTemplate in node.Edges.GroupBy(e => e.Template))
-                {
-                    switch (edgesPerTemplate.Count())
-                    {
-                        case 1: // end point
-                            break;
-                        case 2: // TODO normal road, create segments for tight/broad curves
-                            break;
-                        case 3:
-                            var halfWidth = edgesPerTemplate.Key.RoadWidth / 2;
-
-                            // TODO figure out orientation and endpoints
-                            var segment = new TRoadSegment(
-                                node.Position,
-                                new RoadSegmentEndPoint(node.Position + new Vector3(0, halfWidth, 0)),
-                                new RoadSegmentEndPoint(node.Position + new Vector3(halfWidth, 0, 0)),
-                                new RoadSegmentEndPoint(node.Position + new Vector3(0, -halfWidth, 0)));
-
-                            void Connect(RoadTopologyEdge edge, RoadSegmentEndPoint endPoint, in Vector3 direction)
-                            {
-                                var edgeSegment = edgeSegments[edge];
-                                if (edge.Start.Position == node.Position)
-                                {
-                                    edgeSegment.Start.Position = endPoint.Position;
-                                    edgeSegment.Start.ConnectTo(segment, direction);
-                                    endPoint.ConnectTo(edgeSegment, edge.Start.Position - edge.End.Position);
-                                }
-                                else
-                                {
-                                    edgeSegment.End.Position = endPoint.Position;
-                                    edgeSegment.End.ConnectTo(segment, direction);
-                                    endPoint.ConnectTo(edgeSegment, edge.End.Position - edge.Start.Position);
-                                }
-                            }
-
-                            // TODO consider ordering of edges
-
-                            Connect(edgesPerTemplate.ElementAt(0), segment.Top, Vector3.UnitY);
-                            Connect(edgesPerTemplate.ElementAt(1), segment.Right, Vector3.UnitX);
-                            Connect(edgesPerTemplate.ElementAt(2), segment.Bottom, -Vector3.UnitY);
-
-                            break;
-                    }
-                }
-            }
-
-            foreach (var templateEdges in Edges.GroupBy(e => e.Template))
-            {
-                // Create one network for each connected set of segments of a specific type.
-                var edgesToProcess = new List<AngledRoadSegment>(templateEdges.Select(e => edgeSegments[e]));
-
-                while (edgesToProcess.Count > 0)
-                {
-                    var edgeSegment = edgesToProcess[edgesToProcess.Count - 1];
-                    edgesToProcess.RemoveAt(edgesToProcess.Count - 1);
-
-                    var seenSegments = new List<IRoadSegment>();
-                    seenSegments.Add(edgeSegment);
-
-                    var network = new RoadNetwork(templateEdges.Key);
-                    networks.Add(network);
-
-                    network.AddSegment(edgeSegment);
-
-                    void FollowPath(RoadSegmentEndPoint endPoint)
-                    {
-                        if (endPoint.To == null || seenSegments.Contains(endPoint.To))
-                            return;
-
-                        network.AddSegment(endPoint.To);
-                        seenSegments.Add(endPoint.To);
-
-                        foreach (var nextEndPoint in endPoint.To.EndPoints)
-                        {
-                            FollowPath(nextEndPoint);
-                        }
-                    }
-
-                    foreach (var endPoint in edgeSegment.EndPoints)
-                    {
-                        FollowPath(endPoint);
-                    }
-                }
-            }
-
-            return networks;
-        }
     }
 
     internal sealed class RoadTopologyNode
